+++ annim/src/entities/track.rs
@@ -7,8 +7,8 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
-    pub album_db_id: i64,
-    pub disc_db_id: i64,
+    pub album_db_id: i32,
+    pub disc_db_id: i32,
     pub index: i32,
     pub title: String,
     pub artist: String,
